format and fix latent errors
