+++ components/nimbus/src/lib.rs
@@ -49,33 +49,3 @@ pub use evaluator::TargetingAttributes;
 
 #[cfg(test)]
 mod tests;
-
-cfg_if::cfg_if! {
-    if #[cfg(
-        feature = "stateful-uniffi-bindings"
-    )] {
-        use enrollment::{EnrollmentChangeEvent, EnrollmentChangeEventType};
-        use serde_json::Value;
-
-        impl UniffiCustomTypeConverter for JsonObject {
-            type Builtin = String;
-
-            fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-                let json: Value = serde_json::from_str(&val)?;
-
-                match json.as_object() {
-                    Some(obj) => Ok(obj.clone()),
-                    _ => Err(uniffi::deps::anyhow::anyhow!(
-                        "Unexpected JSON-non-object in the bagging area"
-                    )),
-                }
-            }
-
-            fn from_custom(obj: Self) -> Self::Builtin {
-                serde_json::Value::Object(obj).to_string()
-            }
-        }
-
-        include!(concat!(env!("OUT_DIR"), "/nimbus.uniffi.rs"));
-    }
-}
